#9 move context to context directory
update related import statements. Rename `useDistanceValues` hook to have .js extenstion
