+++ src/App.jsx
@@ -1,8 +1,8 @@
-import { DistanceProvider } from './hooks/useDistanceProvider';
 import './css/App.css';
 import Header from './components/Header';
 import StatsView from './components/StatComponents/StatsView';
 import KilometerInput from './components/KilometerInput';
+import { DistanceProvider } from './context/useDistanceProvider';
 
 function App() {
   return (
